Initial refactor pass
